added google and facebook sign-up

+++ Restro_Dash.py
@@ -27,7 +27,7 @@ session = DBSession()
 def showLogin():
     state=''.join(random.choice(string.ascii_uppercase+string.digits)for x in xrange(32))
     login_session['state']=state
-    return render_template('login.html', STATE = state)
+    return render_template('login.html', STATE = state,login_session=login_session)
 
 
     
@@ -44,7 +44,8 @@ def jsonRes():
 
 @app.route('/')
 def first():
-    results = session.query(Restaurant.name,Restaurant.id).order_by(Restaurant.id.desc()).all()
+    results = session.query(Restaurant).order_by(Restaurant.id.desc()).limit(10).all()
+    creator = -1
     i = 1
     return render_template('homepage.html',results=results, i=i,login_session=login_session)
 
@@ -61,7 +62,7 @@ def newMenuItem(restaurant_id):
 		  flash(newItem.name+" is Created Successfully")
 		  return redirect(url_for('RestaurantMenu', restaurant_id = restaurant_id))
 	   else:
-		  return render_template('newMenuItem.html', restaurant_id = restaurant_id,restaurant=restaurant)
+		  return render_template('newMenuItem.html', restaurant_id = restaurant_id,restaurant=restaurant,login_session=login_session)
 
 @app.route('/restaurants/new', methods=['GET','POST'])
 def newRestaurant():
@@ -74,14 +75,14 @@ def newRestaurant():
             flash(request.form['name']+" already in usage")
             return redirect(url_for('newRestaurant'))
           else:
-            newRes = Restaurant(name = request.form['name'])
+            newRes = Restaurant(name = request.form['name'],user_id= login_session['user_id'])
             user_id= login_session['user_id']
             session.add(newRes)
             session.commit()
             flash(newRes.name+" is Created Successfully")
             return redirect(url_for('first'))
        else:
-          return render_template('newrestaurant.html')
+          return render_template('newrestaurant.html',login_session=login_session)
 
 @app.route('/restaurants/<int:restaurant_id>/<int:MenuID>/edit', methods = ['GET', 'POST'])
 def editMenuItem(restaurant_id, MenuID):
@@ -103,16 +104,21 @@ def editMenuItem(restaurant_id, MenuID):
         flash(editedItem.name+" is edited Successfully")
         return redirect(url_for('RestaurantMenu', restaurant_id = restaurant_id))
     else:
-        return render_template('editmenuitem.html', restaurant_id = restaurant_id, MenuID = MenuID, item = editedItem,restaurant=restaurant)
+        return render_template('editmenuitem.html', restaurant_id = restaurant_id, MenuID = MenuID, item = editedItem,restaurant=restaurant,login_session=login_session)
 
 @app.route('/restaurants/<int:restaurant_id>/delete_restaurant/',methods=['GET','POST'])
 def deleteRestaurant(restaurant_id):
-    if 'username' not in login_session:
-        return redirect(url_for('showLogin'))
     restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
     results = session.query(Restaurant.name,Restaurant.id).order_by(Restaurant.name).all()
+    if 'username' not in login_session:
+        return redirect(url_for('showLogin'))
+    if restaurant.user_id != login_session['user_id']:
+        return '''<script>function myFunction() 
+        {alert('You are not authorized to delete this restaurant. Please create your own restaurant in order to delete');}
+        </script><body onload='myFunction()'>'''
+    
     if request.method=='GET':
-        return render_template('deleterestaurant.html', restaurant = restaurant)
+        return render_template('deleterestaurant.html', restaurant = restaurant,login_session=login_session)
     if request.method=='POST':
         session.delete(restaurant)
         session.commit()
@@ -129,7 +135,11 @@ def MenuItemJson(restaurant_id,menu_id):
 def RestaurantMenu(restaurant_id):
     restaurant=session.query(Restaurant).filter_by(id=restaurant_id).one()
     items=session.query(MenuItem).filter_by(restaurant_id=restaurant.id)
-    return render_template('menu.html', restaurant=restaurant, items=items)
+    creator = getUserInfo(restaurant.user_id)
+    if 'username' not in login_session or creator.id!=login_session['user_id']:
+        return render_template('menupublic.html', restaurant=restaurant, items=items, creator=creator,login_session=login_session)
+    else:
+        return render_template('menu.html', restaurant=restaurant, items=items, creator=creator,login_session=login_session)
 
 @app.route('/restaurants/<int:restaurant_id>/<int:menu_id>/delete_item/',methods=['GET','POST'])
 def deleteMenuItem(restaurant_id, menu_id):
@@ -143,7 +153,85 @@ def deleteMenuItem(restaurant_id, menu_id):
         flash('Menu Item Successfully Deleted')
         return redirect(url_for('RestaurantMenu', restaurant_id=restaurant.id))
     else:
-        return render_template('deleteMenuItem.html', item=itemToDelete , restaurant_id = restaurant.id)
+        return render_template('deleteMenuItem.html', item=itemToDelete , restaurant_id = restaurant.id,login_session=login_session)
+
+@app.route('/fbconnect', methods=['POST'])
+def fbconnect():
+    if request.args.get('state') != login_session['state']:
+        response = make_response(json.dumps('Invalid state parameter.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    access_token = request.data
+    print "access token received %s " % access_token
+
+    app_id = json.loads(open('fb_client_secrets.json', 'r').read())[
+        'web']['app_id']
+    app_secret = json.loads(
+        open('fb_client_secrets.json', 'r').read())['web']['app_secret']
+    url = 'https://graph.facebook.com/oauth/access_token?grant_type=fb_exchange_token&client_id=%s&client_secret=%s&fb_exchange_token=%s' % (
+        app_id, app_secret, access_token)
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[1]
+
+    # Use token to get user info from API
+    userinfo_url = "https://graph.facebook.com/v2.4/me"
+    # strip expire tag from access token
+    token = result.split("&")[0]
+
+
+    url = 'https://graph.facebook.com/v2.4/me?%s&fields=name,id,email' % token
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[1]
+    # print "url sent for API access:%s"% url
+    # print "API JSON result: %s" % result
+    data = json.loads(result)
+    login_session['provider'] = 'facebook'
+    login_session['username'] = data["name"]
+    login_session['email'] = data["email"]
+    login_session['facebook_id'] = data["id"]
+
+    # The token must be stored in the login_session in order to properly logout, let's strip out the information before the equals sign in our token
+    stored_token = token.split("=")[1]
+    login_session['access_token'] = stored_token
+
+    # Get user picture
+    url = 'https://graph.facebook.com/v2.4/me/picture?%s&redirect=0&height=200&width=200' % token
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[1]
+    data = json.loads(result)
+
+    login_session['picture'] = data["data"]["url"]
+
+    # see if user exists
+    user_id = getUserID(login_session['email'])
+    if not user_id:
+        user_id = createUser(login_session)
+    login_session['user_id'] = user_id
+
+    output = ''
+    output += '''
+    <div class="mdl-card mdl-shadow--2dp" style="padding: 12px;">
+    <h3>Welcome,'''
+    output += login_session['username']
+    output += '!</h3>'
+    output += '<img src="'
+    output += login_session['picture']
+    output += ' " style = "width: 100px; height: 100px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"></div>'
+    flash("you are now logged in as %s" % login_session['username'])
+    print "done!"
+    return output
+
+
+@app.route('/fbdisconnect')
+def fbdisconnect():
+    facebook_id = login_session['facebook_id']
+    # The access token must me included to successfully logout
+    access_token = login_session['access_token']
+    url = 'https://graph.facebook.com/%s/permissions?access_token=%s' % (facebook_id,access_token)
+    h = httplib2.Http()
+    result = h.request(url, 'DELETE')[1]
+    return "you have been logged out"
+
 
 @app.route('/gconnect', methods=['POST'])
 def gconnect():
@@ -200,6 +288,7 @@ def gconnect():
         return response
 
     # Store the access token in the session for later use.
+    login_session['provider'] = 'google'
     login_session['credentials'] = credentials.access_token
     login_session['gplus_id'] = gplus_id
 
@@ -217,7 +306,7 @@ def gconnect():
     login_session['user_id']= user_id
     output = ''
     output += '''
-    <div class="mdl-card mdl-shadow--2dp" style="padding: 10px;">
+    <div class="mdl-card mdl-shadow--2dp" style="padding: 12px;">
     <h3>Welcome,'''
     output += login_session['username']
     output += '!</h3>'
@@ -252,6 +341,7 @@ def gdisconnect():
             del login_session['email']
             del login_session['picture']
             
+            flash("Come Back soon here")
             return redirect(url_for('first'))
         else:
         
@@ -260,8 +350,30 @@ def gdisconnect():
             return response
     else:
         return "no user signed-In"
+
+# Disconnect based on provider
+@app.route('/disconnect')
+def disconnect():
+    if 'provider' in login_session:
+        if login_session['provider'] == 'google':
+            gdisconnect()
+            del login_session['credentials']
+        if login_session['provider'] == 'facebook':
+            fbdisconnect()
+            del login_session['facebook_id']
+        del login_session['username']
+        del login_session['email']
+        del login_session['picture']
+        del login_session['user_id']
+        del login_session['provider']
+        flash("You have successfully been logged out.")
+        return redirect(url_for('first'))
+    else:
+        flash("You were not logged in")
+        return redirect(url_for('first'))
+
 def getUserInfo(user_id):
-    user=session.query(User).filter_by(id=user_id).one()
+    user=session.query(User).filter_by(id=user_id).first()
     return user
 
 def createUser(login_session):
@@ -283,5 +395,5 @@ def getUserID(email):
 if __name__=='__main__':
     app.secret_key="you are mad langoor"
     app.debug= True
-    app.run(host="localhost",port=5555) 
+    app.run(host="0.0.0.0",port=5555)
